grngfonttest runs giving dumb rendering

+++ src/TtfTypes.h
@@ -32,8 +32,8 @@ typedef char    byte;
 typedef signed char	int8;
 typedef unsigned short	uint16;
 typedef short 	int16;
-typedef unsigned long	uint32;
-typedef long 	int32;
+typedef unsigned int	uint32;
+typedef int 	int32;
 
 typedef int16	short_frac;
 typedef int32	fixed;
